Refactor and fix atdd-staging
Closes #2004.
