+++ services/src/atdd-staging/.gitignore
@@ -1,3 +1,5 @@
 .idea/ 
 target/
 *.iml
+kilda.properties
+topology.yaml
