+++ services/src/atdd-staging/kilda.properties.example
@@ -2,12 +2,6 @@ northbound.endpoint=http://localhost:8080
 northbound.username=username
 northbound.password=password
 
-topology-engine-rest.endpoint=http://localhost
-topology-engine-rest.username=username
-topology-engine-rest.password=password
-
 floodlight.endpoint=http://localhost:8081
 floodlight.username=username
 floodlight.password=password
-
-lockkeeper.endpoint=http://localhost:5001
