--- services/src/atdd-staging/log4j2.xml.example
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="WARN">
-    <Appenders>
-        <Console name="STDOUT" target="SYSTEM_OUT">
-            <PatternLayout pattern="%d{ISO8601} %-5p %c{1}:%L - %m%n"/>
-        </Console>
-        <File name="File" fileName="logs/request_logs.log">
-            <PatternLayout pattern="%d{ISO8601} %-5p %c{1}:%L - %m%n"/>
-        </File>
-    </Appenders>
-    <Loggers>
-        <Logger name="org.openkilda.testing.tools.LoggingRequestInterceptor" level="DEBUG" additivity="false">
-            <AppenderRef ref="File"/>
-        </Logger>
-        <Root level="INFO">
-            <AppenderRef ref="STDOUT"/>
-        </Root>
-    </Loggers>
-</Configuration>
