+++ services/src/atdd-staging/src/main/java/org/openkilda/atdd/staging/helpers/DefaultFlowsChecker.java
@@ -121,5 +121,4 @@ public final class DefaultFlowsChecker {
 
     private DefaultFlowsChecker() {
     }
-
 }
