+++ services/src/atdd-staging/src/main/java/org/openkilda/atdd/staging/helpers/FlowSet.java
@@ -35,7 +35,7 @@ import java.util.Optional;
 import java.util.Set;
 
 /**
- * A builder for flows ({@link FlowPayload}) in nonoverlapped VLANs.
+ * A builder for flows ({@link FlowPayload}) in non-overlapped VLANs.
  */
 public class FlowSet {
 
@@ -44,19 +44,6 @@ public class FlowSet {
 
     public Set<FlowPayload> getFlows() {
         return unmodifiableSet(flows);
-    }
-
-    /**
-     * Returns an unallocated vlan. The returned vlan is immediately added to the list of allocated vlans
-     */
-    public int allocateVlan() {
-        RangeSet<Integer> availableVlansRange = TreeRangeSet.create();
-        availableVlansRange.removeAll(allocatedVlans);
-        Integer vlan = availableVlansRange.asRanges().stream()
-                .flatMap(range -> ContiguousSet.create(range, DiscreteDomain.integers()).stream())
-                .findFirst().get();
-        allocatedVlans.add(Range.singleton(vlan));
-        return vlan;
     }
 
     public void addFlow(String flowId, Switch srcSwitch, Switch destSwitch) {
@@ -262,4 +249,3 @@ public class FlowSet {
         }
     }
 }
-
