+++ services/src/atdd-staging/src/main/resources/cucumber.properties
@@ -1,3 +1,2 @@
-# This properties file defines default Cucumber options.
-cucumber.options=--glue org.openkilda.atdd.staging.steps --plugin pretty --plugin json:cucumber-reports/report.json --plugin html:cucumber-reports/html classpath:features
-
+# This properties file defines Cucumber options.
+cucumber.options=--glue org.openkilda.atdd.staging.steps --plugin pretty classpath:features
