--- services/src/atdd-staging/src/test/resources/cucumber.properties
@@ -1,1 +1,0 @@
-# This properties file defines Cucumber options for Unit testing. They override the options of the main sources.
